Rename RoutingTripPattern to V2
