+++ src/main/java/org/opentripplanner/transit/model/trip/ScheduledTripPattern.java
@@ -2,11 +2,12 @@ package org.opentripplanner.transit.model.trip;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.opentripplanner.transit.model.network.RoutingTripPatternV2;
 
 /**
  * TODO RTM - THIS IS A PLACEHOLDER for {@link org.opentripplanner.model.TripPattern}
  */
 public class ScheduledTripPattern {
 
-  private final List<RoutingTripPattern> variants = new ArrayList<>();
+  private final List<RoutingTripPatternV2> variants = new ArrayList<>();
 }
